# Clean `test_{runner,scalafmt_helper}`, add diagnostic env vars, and improve `test_cross_build` performance