--- a/test/shell/test_cross_build.sh
+++ b/test/shell/test_cross_build.sh
@@ -9,11 +9,11 @@ cd test_cross_build
 
 function test_cross_build() {
   bazel test //...
-  bazel clean
-  bazel shutdown;
 }
 
 function test_scalafmt() {
+  bazel build //scalafmt/...
+
   run_formatting scalafmt binary2 binary2
   run_formatting scalafmt binary3 binary3
   run_formatting scalafmt library2 library2
@@ -24,3 +24,10 @@ function test_scalafmt() {
 
 $runner test_cross_build
 $runner test_scalafmt
+
+# `bazel shutdown` used to be in `test_cross_build`, after a `bazel clean`.
+# However, the protobuf library tends to rebuild frequently. Not cleaning and
+# postponing the shutdown to the end of the script helps avoid rebuilding as
+# often, speeding up the tests. It also potentially speeds up debugging by
+# preserving the workspace state when a test fails.
+bazel shutdown